fix: remove all previous bot messages
